TOWARDS A RAG SYSTEM
This is a end-of-day commit and does *not* mark a step where the
code is actually working.

At this point there is a retriever and a reranking model. What is still
missing is a working query transformation model. The base code for the
LLM-based answer generation is there.

At this point, none of the code in this file have been tested yet and
bugs are expected.