MINOR FIXES AND CHANGES
The code was streamlined in order to increase modularity. The chunks
and scores are now merely updated during reranking such that the
code works even in reranking is not used.

Further, some minor bugs were fixed.

AT THIS POINT the code works from a technical perspective, i.e.
it does not crash and does not throw errors under the following conditions:
- Chroma is used as a vector store
- Query rewriting is used in query transformation
- Reranking is not used at all

All these additional features are yet to be tested or even implemented.